tests/wpt/web-platform-tests/web-animations/responsive/strokeDasharray.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id='container'>
    <div id='element'></div>
</div>
<script>

var container = document.getElementById('container');
var element = document.getElementById('element');

test(function() {
    container.style.fontSize = '10px';

    var keyframes = [
        {strokeDasharray: '10em 10em'},
        {strokeDasharray: '10em 10em'}
    ];
    var player = element.animate(keyframes, 10);
    player.pause();
    player.currentTime = 5;

    var strokeDasharray = getComputedStyle(element).strokeDasharray;
    container.style.fontSize = '20px';
    assert_not_equals(getComputedStyle(element).strokeDasharray, strokeDasharray);
}, 'strokeDasharray responsive to style changes');

</script>
